requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.5
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.1.31
chardet==5.2.0
crispy-bootstrap5==2024.2
Django==5.0.2
django-crispy-forms==2.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
magic-filter==1.0.12
multidict==6.4.3
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.0
pillow==10.2.0
propcache==0.3.1
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot==22.0
pytz==2025.2
reportlab==4.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
yarl==1.19.0

